Rename read_and_write_file folder to io
