+++ src/internals/io/read_write_file.h
@@ -1,0 +1,6 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+char* read_file(const char* file_in);
+FILE* create_out_file(const char* file_in);
