+++ tests/read_write_file.c
@@ -1,4 +1,4 @@
-#include <../src/internals/read_and_write_file/read_write_file.h>
+#include <../src/internals/io/read_write_file.h>
 
 char* read_file(const char* file_in)
 {
